requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sendgrid
itsdangerous
